Move Home route to kapture-backend
