+++ ui/src/implementations/kapture-backend/App.js
@@ -4,7 +4,7 @@ import { BrowserRouter, Route, Switch, Redirect } from 'react-router-dom';
 
 import { store } from '../../store/store';
 import { Header } from '../../components/app/Header';
-import { Home } from '../../routes/home';
+import { Home } from './routes/home';
 import { Activities } from '../../routes/activities';
 import styles from './App.module.css';
