+++ ui/src/implementations/kapture-backend/components/index.js
@@ -1,0 +1,1 @@
+export { ActivitySearch } from '../../../components/activity/ActivitySearch';
